fix(quality): address PR review findings
- harden CI token permissions and action pinning
- remove cwd-dependent test setup and ambiguous fixture imports
- cover multi-primary-key commit mapping and isolate smoke tests
- tighten integration cleanup and external test markers
