+++ .workflow/quality-program/reports/production-change-ledger.md
@@ -5,3 +5,6 @@
 | G2 | `hugegraph-python-client/src/pyhugegraph/utils/util.py` | Preserve backend error envelope details for non-404 HTTP errors and prefer server `message` over `exception`. | `uv run pytest hugegraph-python-client/src/tests/api/test_response_validation.py -q` | Response validation contract gap for 500 backend envelopes. | Low; narrows error wrapping to preserve parsed server details while keeping 404 mapping. |
 | G3 | `hugegraph-llm/src/hugegraph_llm/operators/hugegraph_op/commit_to_hugegraph.py` | Map primary-key `label:value` edge endpoints to server-created VIDs and raise explicit vertex creation errors before accessing `.id`. | `uv run pytest hugegraph-llm/src/tests/operators/hugegraph_op/test_commit_to_hugegraph.py::TestCommit2Graph::test_load_into_graph_raises_explicit_error_when_vertex_creation_fails -q`; `HUGEGRAPH_REQUIRED=true uv run pytest hugegraph-llm/src/tests/integration/test_hugegraph_boundary.py -v --tb=short` | LLM HugeGraph write-boundary contract gap for primary-key vertices and failure-surface clarity. | Low; scoped to failed vertex creation and endpoint mapping after successful creation. |
 | G4 | `hugegraph-llm/src/hugegraph_llm/api/rag_api.py` | Map `/config/llm` to existing `llm_settings.chat_llm_type` instead of nonexistent `llm_settings.llm_type`. | `uv run pytest hugegraph-llm/src/tests/api -m "unit or contract" -v --tb=short` | API public-surface field mapping contract gap. | Low; aligns route with existing config schema and leaves callback arguments unchanged. |
+| PR68 review | `hugegraph-llm/src/hugegraph_llm/operators/hugegraph_op/commit_to_hugegraph.py` | Extend primary-key endpoint mapping to multi-key vertex labels by joining primary key values with `!`. | `uv run pytest hugegraph-llm/src/tests/operators/hugegraph_op/test_commit_to_hugegraph_load_into_graph.py -q` | Review-identified edge creation gap for multiple primary-key vertex IDs. | Low; only broadens fallback mapping for existing primary-key schemas. |
+| PR68 review | `hugegraph-llm/src/hugegraph_llm/config/models/base_prompt_config.py` | Resolve prompt config YAML from the package project root instead of process cwd. | `uv run pytest hugegraph-llm/src/tests/config/test_config.py -q` | Avoid import-time cwd mutation in test harness and make config path deterministic. | Medium-low; preserves the same file path when run from module root and removes cwd guard. |
+| PR68 review | `hugegraph-llm/src/hugegraph_llm/config/models/base_config.py` | Resolve `.env` from the package project root instead of process cwd. | `uv run pytest hugegraph-llm/src/tests/config/test_config.py -q` | Avoid writing or reading a caller cwd `.env` when tests import config from repo root. | Medium-low; preserves the same `.env` file when run from module root and removes cwd coupling. |
