+++ hugegraph-llm/src/hugegraph_llm/config/models/base_config.py
@@ -21,10 +21,11 @@ import os
 from dotenv import dotenv_values, set_key
 from pydantic_settings import BaseSettings
 
+from hugegraph_llm.utils.anchor import get_project_root
 from hugegraph_llm.utils.log import log
 
 dir_name = os.path.dirname
-env_path = os.path.join(os.getcwd(), ".env")  # Load .env from the current working directory
+env_path = os.path.join(get_project_root(), ".env")
 
 
 class BaseConfig(BaseSettings):
