+++ hugegraph-llm/src/tests/config/test_config.py
@@ -17,6 +17,7 @@
 
 
 import unittest
+from pathlib import Path
 
 import pytest
 
@@ -31,3 +32,15 @@ class TestConfig(unittest.TestCase):
 
         nltk.data.path.append(resource_path)
         nltk.data.find("corpora/stopwords")
+
+    def test_prompt_yaml_path_is_project_root_independent(self):
+        from hugegraph_llm.config.models import base_prompt_config
+
+        expected = Path(__file__).resolve().parents[2] / "hugegraph_llm" / "resources" / "demo" / "config_prompt.yaml"
+        self.assertEqual(Path(base_prompt_config.yaml_file_path), expected)
+
+    def test_env_path_is_project_root_independent(self):
+        from hugegraph_llm.config.models import base_config
+
+        expected = Path(__file__).resolve().parents[3] / ".env"
+        self.assertEqual(Path(base_config.env_path), expected)
