+++ hugegraph-llm/src/tests/operators/llm_op/test_gremlin_generate.py
@@ -22,10 +22,10 @@ import unittest
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
+from fixtures.fake_llm import FakeLLM
 
 from hugegraph_llm.models.llms.base import BaseLLM
 from hugegraph_llm.operators.llm_op.gremlin_generate import GremlinGenerateSynthesize
-from tests.fixtures.fake_llm import FakeLLM
 
 pytestmark = pytest.mark.contract
