+++ hugegraph-llm/src/tests/operators/llm_op/test_keyword_extract.py
@@ -21,10 +21,10 @@ import unittest
 from unittest.mock import MagicMock, patch
 
 import pytest
+from fixtures.fake_llm import FakeLLM
 
 from hugegraph_llm.models.llms.base import BaseLLM
 from hugegraph_llm.operators.llm_op.keyword_extract import KeywordExtract
-from tests.fixtures.fake_llm import FakeLLM
 
 pytestmark = pytest.mark.contract
