+++ hugegraph-python-client/src/tests/api/test_schema.py
@@ -16,6 +16,7 @@
 # under the License.
 
 import unittest
+from contextlib import suppress
 
 import pytest
 
@@ -87,21 +88,36 @@ class TestSchemaManager(unittest.TestCase):
 
 def test_schema_create_and_fetch_property_vertex_edge_index(client_utils):
     schema = client_utils.schema
-    schema.propertyKey("quality_name").asText().ifNotExist().create()
-    schema.propertyKey("quality_score").asInt().ifNotExist().create()
-    schema.vertexLabel("quality_person").properties("quality_name", "quality_score").primaryKeys(
-        "quality_name"
-    ).ifNotExist().create()
-    schema.edgeLabel("quality_knows").sourceLabel("quality_person").targetLabel("quality_person").ifNotExist().create()
-    schema.indexLabel("quality_person_by_score").onV("quality_person").by("quality_score").range().ifNotExist().create()
+    try:
+        schema.propertyKey("quality_name").asText().ifNotExist().create()
+        schema.propertyKey("quality_score").asInt().ifNotExist().create()
+        schema.vertexLabel("quality_person").properties("quality_name", "quality_score").primaryKeys(
+            "quality_name"
+        ).ifNotExist().create()
+        schema.edgeLabel("quality_knows").sourceLabel("quality_person").targetLabel(
+            "quality_person"
+        ).ifNotExist().create()
+        schema.indexLabel("quality_person_by_score").onV("quality_person").by(
+            "quality_score"
+        ).range().ifNotExist().create()
 
-    full_schema = schema.getSchema()
+        full_schema = schema.getSchema()
 
-    assert "propertykeys" in full_schema
-    assert "vertexlabels" in full_schema
-    assert "edgelabels" in full_schema
-    assert "indexlabels" in full_schema
-    assert schema.getPropertyKey("quality_name").name == "quality_name"
-    assert schema.getVertexLabel("quality_person").name == "quality_person"
-    assert schema.getEdgeLabel("quality_knows").name == "quality_knows"
-    assert schema.getIndexLabel("quality_person_by_score").name == "quality_person_by_score"
+        assert "propertykeys" in full_schema
+        assert "vertexlabels" in full_schema
+        assert "edgelabels" in full_schema
+        assert "indexlabels" in full_schema
+        assert schema.getPropertyKey("quality_name").name == "quality_name"
+        assert schema.getVertexLabel("quality_person").name == "quality_person"
+        assert schema.getEdgeLabel("quality_knows").name == "quality_knows"
+        assert schema.getIndexLabel("quality_person_by_score").name == "quality_person_by_score"
+    finally:
+        for remove in (
+            lambda: schema.indexLabel("quality_person_by_score").remove(),
+            lambda: schema.edgeLabel("quality_knows").remove(),
+            lambda: schema.vertexLabel("quality_person").remove(),
+            lambda: schema.propertyKey("quality_score").remove(),
+            lambda: schema.propertyKey("quality_name").remove(),
+        ):
+            with suppress(Exception):
+                remove()
